refactor: Token生成部分を別クラスに切り出し
